Rename: square.download_manifests -> manio.download
